# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API do sistema GuardPoint (controle de rondas/postos).",
        "title": "GuardPoint API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/alertas": {
            "get": {
                "description": "Para supervisores, os alertas sao automaticamente filtrados pelos postos aos quais estao vinculados (via posto_supervisores). Admins veem todos os alertas. Os filtros de query sao adicionais — posto_id pode ser usado para refinar a busca.",
                "tags": [
                    "alertas"
                ],
                "summary": "Lista alertas com filtros e paginacao (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status do alerta (aberto, reconhecido, encerrado)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipo do alerta (atraso, no_show, senha_emergencia, senha_customizada, sabotagem)",
                        "name": "tipo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID do turno (uuid)",
                        "name": "turno_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID do posto (uuid)",
                        "name": "posto_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite de itens (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset da paginacao",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertaListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/estatisticas": {
            "get": {
                "tags": [
                    "alertas"
                ],
                "summary": "Estatisticas agregadas de alertas (admin/supervisor)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertStatistics"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/{id}/encerrar": {
            "put": {
                "tags": [
                    "alertas"
                ],
                "summary": "Encerra um alerta (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do alerta (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alertas/{id}/reconhecer": {
            "put": {
                "tags": [
                    "alertas"
                ],
                "summary": "Reconhece um alerta aberto (admin/supervisor)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do alerta (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_guardpoint_guardpoint-server_internal_model.AlertStatistics": {
            "type": "object",
            "properties": {
                "por_hora": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertaPorHora"
                    }
                },
                "por_tipo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.AlertaPorTipo"
                    }
                },
                "total_abertos": {
                    "type": "integer",
                    "example": 5
                },
                "total_encerrados": {
                    "type": "integer",
                    "example": 12
                },
                "total_reconhecidos": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.Alerta": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "empresa_id": {
                    "type": "string",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "mensagem": {
                    "type": "string",
                    "example": "Atraso de 5 minutos detectado no turno."
                },
                "nivel": {
                    "type": "integer",
                    "example": 1
                },
                "posto_id": {
                    "type": "string",
                    "example": "880e8400-e29b-41d4-a716-446655440000"
                },
                "resolvido_em": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "aberto"
                },
                "tipo": {
                    "type": "string",
                    "example": "atraso"
                },
                "turno_id": {
                    "type": "string",
                    "example": "770e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.AlertaListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_guardpoint_guardpoint-server_internal_model.Alerta"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.AlertaPorHora": {
            "type": "object",
            "properties": {
                "hora": {
                    "type": "string",
                    "example": "08:00"
                },
                "quantidade": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
        "github_com_guardpoint_guardpoint-server_internal_model.AlertaPorTipo": {
            "type": "object",
            "properties": {
                "quantidade": {
                    "type": "integer",
                    "example": 8
                },
                "tipo": {
                    "type": "string",
                    "example": "atraso"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensagem descritiva do erro"
                }
            }
        },
...
        "github_com_guardpoint_guardpoint-server_internal_model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}